refactor and add approve_all event data
