+++ contracts/nft/src/execute/approve_all.rs
@@ -2,6 +2,7 @@ use cosmwasm_std::{Addr, DepsMut, Env, MessageInfo, Response};
 use cw_utils::Expiration;
 
 use crate::core::error::ContractError;
+use crate::events::approve_all::EventApproveAll;
 use crate::storage::operators::OPERATORS;
 use crate::util::action::{Action, ActionType};
 
@@ -21,5 +22,13 @@ pub fn handle(
     // set the operator for us
     OPERATORS.save(deps.storage, (&info.sender, &operator), &expires)?;
 
-    Ok(Response::default().set_action(ActionType::ApproveAll))
+    Ok(Response::default()
+        .set_action(ActionType::ApproveAll)
+        .add_event(
+            EventApproveAll {
+                operator,
+                sender: info.sender,
+            }
+            .into(),
+        ))
 }
